refactor : enum 클래스 구현에 따른 코드 변경.
세부 사항
- 모호한 필드 네이밍 수정.
- getter 메소드 제거.
